api/requirements.txt
fastapi
openai
python-dotenv
pydantic
uvicorn
